outside range

+++ src/stable/interfaces/IFeeConfiguration.sol
@@ -4,18 +4,16 @@ pragma solidity ^0.8.0;
 import {PoolId} from "@uniswap/v4-core/src/types/PoolId.sol";
 
 struct FeeConfig {
-    // TODO: natspec
-    uint256 k;
-    uint256 logK;
-    uint24 optimalFeeRate;
-    uint160 referenceSqrtPriceX96;
+    uint256 k; // Decay rate per block; controls how fast fees decrease toward target
+    uint256 logK; // Used for efficient decay calculation over many blocks
+    uint24 optimalFeeRate; // Optimal range width in 1e6 precision; inside = consistent buy/sell prices, outside = flexible
+    uint160 referenceSqrtPriceX96; // Reference sqrt price; optimal range centered around this
 }
 
 struct FeeState {
-    // TODO: natspec
-    uint256 previousFee;
-    uint160 previousSqrtAmmPriceX96;
-    uint256 blockNumber;
+    uint40 previousFee; // Last flexible fee charged in 1e12 precision; used for exponential decay calculation
+    uint160 previousSqrtAmmPriceX96; // AMM sqrt price at last swap; used to detect price movement direction
+    uint256 blockNumber; // Block when fee was last updated; determines decay based on blocks elapsed
 }
 
 /// @notice Interface for the FeeConfiguration
